build: cancel CI builds after a PR was updated
Configure GitHub Actions Workflow to cancel any pending or in-progress
runs of our CI build after the pull request being tested has been
modified. We are interested in the result of the new build that will
be started for the updated version.

Preserve the old behavior for the `main` branch, where we want to
build every commit even when new commits were added before the CI
finished.

Signed-off-by: Miroslav Bajtoš <saturn@bajtos.net>