feat: derive PASSPHRASE for Core from the wallet
Create a PASSPHRASE for Station Core to let it encrypt the station
identity stored on the filesystem.

The passphrase is created by signing a constant message using the
private wallet key that's unique to each Station Desktop instance.
This way the passphrase is unique, cannot be guessed without the
knowledge of the wallet private key, and stays the same accross
Station Desktop restarts.

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>